Let the timer actually run, and give replay speeds slower than real time
The last engine acceptance criterion: scheduling rather than execution. Every other replay test drives the engine
synchronously, which proves the simulation is right and says nothing about the timer.

Seven tests with the timer firing. Play advances on its own, pause stops it and it stays stopped however long you
wait, resume picks up from where it paused, reaching the end leaves the transport paused at the end with play no
longer offered, and two timer-driven playbacks of the same span at 30× and 4× reach the same fingerprint.

Two things worth reading rather than skimming.

The 0.25× and 0.5× speeds you asked for did not exist — the live simulation offers integer 1…30, so I added them
rather than quietly testing something else. ReplayClock now has its own `Double` speed set, [0.25, 0.5, 1, 2, 4,
10, 30], separate from live's on purpose: watching an approach at a quarter speed is a review tool, and giving a
*live* exercise fractional speed would mean it ran slower than the clock a trainee is being timed against.

These tests take real time, so they are written not to be flaky. Waits are as short as the fastest speed allows,
and the assertions are about direction — "it advanced", "it did not advance" — because how many times a timer
fires in 300 ms is a property of the machine, not of the engine. Asserting an exact tick count would be asserting
the CI runner's mood. The speed-comparison test waits until a position is *reached* rather than for a fixed
duration, so a slow machine takes longer instead of producing a different answer, and then settles to the exact
target before fingerprinting — a timer may overshoot by one fire, and that is scheduling, not divergence.

`letTimersFire(for:)` spins the run loop through the main queue rather than sleeping. `Thread.sleep` would block
the run loop and no timer would fire, which would make every test in this file pass for the wrong reason.

Engine acceptance now covers all of it: deterministic replay against recorded fingerprints, pause and resume,
variable speed with identical outcomes, forward and backward seeking, continue from an arbitrary point, branch
creation, self-sufficient branch replay, no divergence after continuing, and measured seek latency.

Next is Item 9, the UI, which should hold no replay state of its own.

+++ VectraPro/Commands/ReplayEngine.swift
@@ -251,7 +251,7 @@ final class ReplayEngine {
     ///
     /// Only the interval changes. Folding speed into the step size is how a replay stops reaching the same
     /// state as the recording.
-    func setSpeed(_ speed: Int) {
+    func setSpeed(_ speed: Double) {
         clock.setSpeed(speed)
         if clock.isRunning { schedule() }
     }
